refactor: change directory of Hono client helper
