+++ app/routes/staff/orders/new/-components/$orderRegistrationForm.tsx
@@ -9,7 +9,7 @@ import Chip from "../../../../../components/ui/chip"
 import ChipButton from "../../../../../components/ui/chipButton"
 import Input from "../../../../../components/ui/input"
 import Label from "../../../../../components/ui/label"
-import { createHonoClient } from "../../../../../helpers/hono/hono-client"
+import { createHonoClient } from "../../../../../helpers/api/hono-client"
 import type { OrderRegistrationPageData } from "../../../../../usecases/getOrderRegistrationPageData"
 import type { RegisterOrderParams } from "../../../../../usecases/registerOrder"
 import { formatCurrencyJPY } from "../../../../../utils/money"
